move: components 디렉토리 구조 수정으로 인한 파일 이동
