+++ src/components/PageStyledComponents/admin-ApplicationOpenPage/setPeople/index.tsx
